Fix clang warnings.
When building Material.cc with clang, it warns that the loop
construct used is a reference, but a copy is going to be made
anyway because of a const/non-const mismatch.  Fix this by
adding const to the loop variable type, which removes the
warning and should be more efficient.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>