Adds a container class for  numbers.
This commit adds a container class for values also. In particular it implements the behaviour for the class `TimeVaryingVolumetricGrid<T, V, InMemorySession<T, double>>` . This class serves as a lookup table for time varying items. To construct a `TimeVaryingVolumetricGrid` it is advised to use the `InMemoryTimeVaryingVolumetricGridFactory` to construct the actual grid from a given dataset. Do not that this dataset should fit in memory.

To progress along the time axis use the `InMemorySession` object. This can be acquired using the `CreateSession` function. To set the time to be queried, step the `InMemorySession`
using the `StepTo` function. Finally to look up a value use the `LookUp` function.

## Example

```
 InMemoryTimeVaryingVolumetricGridFactory<double, double> gridFactory;

  for (double t = 0; t < 1; t+=0.2)
  {
    for (double x = 0; x < 1; x+=0.5)
    {
      for (double y = 0; y < 1; y+=0.5)
      {
        for (double z = 0; z < 1; z+=0.5)
        {
          gridFactory.AddPoint(t, Vector3d{x,y,z}, t);
        }
      }
    }
  }

 /// Construct the grid
 auto grid = gridFactory.Build();

 /// Get the session cursor
 auto session = grid.CreateSession();

 /// Step to the correct time
 auto new_sess = grid.StepTo(session, 0.5);

 /// Retrieve value from given location
 auto  val = grid.LookUp(new_sess.value(), Vector3d{0.5, 0.5, 0.5});
```

## Other Changes in this commit
This commit also makes some changes to `VolumetricGridLookUpField`. In particular the changes allow
 `VolumetricGridLookUpField` to use custom indices and support looking up in larger
index slices. There is also a new constructor for it to support the `TimeVaryingVolumetricGridLookupField`. This shouldn;t really be a problem as the `VolumetricGridField` is only available on the main branch and not in any collection yet.

## TODO
There is still a fair deal of work to do by ways of documentation and test coverage. I would also like to change the behaviour of what happens when we go out of bounds of the grid.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>