Adds a validity check for Sessions created using the `TimeVaryingVolumetricGrid`
This is needed to resolve a bug in https://github.com/gazebosim/gz-sim/pull/1842#pullrequestreview-1596324924. In particular part of the issue is poor API design. Ideally `CreateSession` would return an `Option<Session>`, however it returns a `Session` even if it is not possible to create the session (for instance if it is too far ahead in time or an empty grid). Hence downstream users need a way to verify session validity.

Signed-off-by: Arjo Chakravarty <arjoc@intrinsic.ai>