Replace CMake Python variables with new ones from FindPython3 module (#402)
Replaces the old variables with the new ones used by the FindPython3 module. Right now IgnPython cannot be used as it only searches for the interpreter but pybind11 also needs the development components, thus it provides a workaround. The Win32 PYTHON_LIBRARIES (PYTHON_DEBUG_LIBRARIES) variable was also removed as there is no equivalent in the new CMake module for getting the debug libraries.

---------

Signed-off-by: Bi0T1N <Bi0T1N@users.noreply.github.com>
Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>
Co-authored-by: Bi0T1N <Bi0T1N@users.noreply.github.com>
Co-authored-by: Jose Luis Rivero <jrivero@osrfoundation.org>
Co-authored-by: Steve Peters <scpeters@openrobotics.org>
Co-authored-by: Addisu Z. Taddese <addisu@openrobotics.org>