Encapsulate std::cerr under a single point of control
When integrating this library into a larger application, downstream
code doesn't always want to spew onto cerr. To allow customization,
we'll refactor to encapsulate the destination inside of a function.
Future work can extent this with a mechanism to redirect it (e.g.,
to send it to spdlog instead of cerr).

Signed-off-by: Jeremy Nimmer <jeremy.nimmer@tri.global>